Moving some code around to reduce the import-depth in unit tests
