+++ src/layout/Accordion/index.tsx
@@ -5,6 +5,7 @@ import { AccordionDef } from 'src/layout/Accordion/config.def.generated';
 import { AccordionHierarchyGenerator } from 'src/layout/Accordion/hierarchy';
 import { SummaryAccordionComponent } from 'src/layout/Accordion/SummaryAccordion';
 import type { PropsFromGenericComponent } from 'src/layout';
+import type { SummaryRendererProps } from 'src/layout/LayoutComponent';
 import type { ComponentHierarchyGenerator } from 'src/utils/layout/HierarchyGenerator';
 
 export class Accordion extends AccordionDef {
@@ -18,7 +19,7 @@ export class Accordion extends AccordionDef {
     return this._hierarchyGenerator;
   }
 
-  renderSummary(props): JSX.Element | null {
+  renderSummary(props: SummaryRendererProps<'Accordion'>): React.JSX.Element | null {
     return <SummaryAccordionComponent {...props} />;
   }
